+++ docs/release_notes.md
@@ -14,6 +14,7 @@
 
 - [Orchestration] Added `GEMINI_3_1_FLASH_LITE` to model list in `OrchestrationAiModel`.
 - [Batch] Added a new `BatchesApi` client to access the [Batch Service API](https://github.tools.sap/AI/llm-batch-service).
+- Added new OpenAI GPT-5.5 model: `OpenAiModel.GPT_55`and `OrchestrationAiModel.GPT_55`.
 
 ### 📈 Improvements